requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
anyio==4.0.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
courlan==1.3.0
cron-descriptor==1.4.0
cryptography==43.0.0
cssselect==1.2.0
dateparser==1.2.0
distro==1.9.0
Django==3.2.18
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-filter==23.2
django-phonenumber-field==7.1.0
django-simpleui==2023.3.1
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.4.2
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.1.3
filelock==3.12.3
frozenlist==1.4.1
fsspec==2023.6.0
goose3==3.1.19
greenlet==3.0.1
grpcio==1.57.0
grpcio-tools==1.57.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
htmldate==1.8.1
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
joblib==1.3.2
jusText==3.0.1
kombu==5.3.2
langdetect==1.0.9
lit==16.0.6
lxml==5.2.2
lxml_html_clean==0.2.0
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.5
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai==0.28.0
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.4
Pillow==10.0.0
portalocker==2.7.0
prompt-toolkit==3.0.39
protobuf==4.24.2
pyahocorasick==2.1.0
pycparser==2.22
pydantic==2.3.0
pydantic_core==2.6.3
PyJWT==2.7.0
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
PyMySQL==1.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qdrant-client==1.10.1
redis==5.0.1
regex==2023.8.8
requests==2.31.0
retrying==1.3.4
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
simpleui==4.0.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
sympy==1.12
threadpoolctl==3.2.0
tld==0.13
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
tqdm-joblib==0.0.3
trafilatura==1.12.0
transformers==4.32.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.4
